Refactored DTOs

+++ pojo/src/main/java/com/stanzaliving/boq_service/dto/LabelValueCountDto.java
@@ -19,5 +19,5 @@ public class LabelValueCountDto<T> {
     @NotBlank
     private T value;
 
-    private BigInteger count;
+    private long count;
 }
